Padronização dos códigos
