+++ 4food/src/actions/user.js
@@ -1,5 +1,6 @@
 import axios from "axios";
-
+import { push } from "connected-react-router";
+import { routes } from "../containers/Router";
 
 const baseURL = "https://us-central1-missao-newton.cloudfunctions.net/FourFoodB"
 
@@ -12,7 +13,7 @@ const loginUser = (login) => ({
 })
 
 
-export const postLoginUser = (email, password) => async (dispatch) =>{
+export const postLoginUser = (email, password) => async (dispatch) => {
 
     const newLoginUser = {
         email,
@@ -27,5 +28,24 @@ export const postLoginUser = (email, password) => async (dispatch) =>{
         window.alert("Login realizado com sucesso!!");
     }catch(error){
         window.alert("Email ou senha incorreta.")
+    }
+}
+
+
+export const signUp = (username, email, cpf, password, confirmPassword) => async (dispatch) => {
+    const signUpInfo = {
+        username: username,
+        email: email,
+        cpf: cpf,
+        password: password,
+        confirmPassword: confirmPassword,
+    };
+
+    try {
+        const response = await axios.post (`${baseURL}/signup`, signUpInfo)
+        window.localStorage.getItem("token", response.data.token);
+        dispatch(push(routes.root))
+    } catch (error) {
+        window.alert("Não foi possível realizar o cadastro")
     }
 }
