+++ 4food/src/resources/logo-future-eats-invert.png
